Refactor memory system
