+++ camel/memories/blocks/__init__.py
@@ -12,22 +12,10 @@
 # limitations under the License.
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 
-from .records import MemoryRecord, ContextRecord
-from .base import MemoryBlock, AgentMemory, BaseContextCreator
-from .context_creators.score_based import ScoreBasedContextCreator
-from .blocks.chat_history_block import ChatHistoryBlock
-from .blocks.vectordb_block import VectorDBBlock
-from .example_agent_memories import VectorDBMemory, ChatHistoryMemory
+from .chat_history_block import ChatHistoryBlock
+from .vectordb_block import VectorDBBlock
 
 __all__ = [
-    'MemoryRecord',
-    'ContextRecord',
-    'MemoryBlock',
-    "AgentMemory",
-    'BaseContextCreator',
-    'ScoreBasedContextCreator',
-    'ChatHistoryMemory',
-    'VectorDBMemory',
     'ChatHistoryBlock',
     'VectorDBBlock',
 ]
